+++ CHANGES.rst
@@ -2,6 +2,8 @@ Next release
 ---------------
 - [Bug] :issue:`109`: Fix automated session closure handled by python garbage collection
 - [Bug] :issue:`120`: Fix get_remote_session not respecting 'timeout' parameter
+- [Bug] :issue:`139`: Fix run_cmd raising AuthenticationException if no agent is running
+
 
 1.6.3 (03/12/2020)
 ------------------